Move set proxy into its own file
